requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
openai==1.43.0
vosk==0.3.45
# numpy==1.26.4
soundfile==0.12.1
pyttsx3==2.90
pydantic==2.9.2
